backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
langgraph
langchain
langchain-groq
python-dotenv
